+++ qpanel/freeswitch.py
@@ -3,7 +3,7 @@
 #
 # Class Qpanel for Freeswitch
 #
-# Copyright (C) 2015-2019 Rodrigo Ramírez Norambuena <a@rodrigoramirez.com>
+# Copyright (C) 2015-2020 Rodrigo Ramírez Norambuena <a@rodrigoramirez.com>
 #
 
 
@@ -85,7 +85,7 @@ class Freeswitch:
 
         for l in tmp[1:-1]:
             line = l.split(delimiter)
-            if len(line) < len(header):
+            if len(line) != len(header):
                 continue
             i = 0
             tmp_dict = {}